dev: move to test_utils for eth-rpc
