+++ crates/eth-rpc/src/lib.rs
@@ -7,6 +7,7 @@ pub mod api;
 pub mod config;
 pub mod rpc;
 pub mod servers;
+pub mod test_utils;
 
 use eyre::Result;
 use jsonrpsee::server::{ServerBuilder, ServerHandle};
